eeee

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long int d;
 	long long int g;
 
-	printf("Size of an int: %lu byte(s)\n", (unsigned long) sizeof(i));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long) sizeof(f));
-	printf("Size of a char: %lu byte(S)\n", (unsigned long) sizeof(c));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long) sizeof(d));
-	printf("Size of a long long int: %ln byte(s)\n", (unsigned long) sizeof(g));
+	printf("Size of an int: %lu byte(s)\n",  sizeof(i));
+	printf("Size of a float: %lu byte(s)\n",  sizeof(f));
+	printf("Size of a char: %lu byte(S)\n",  sizeof(c));
+	printf("Size of a long int: %lu byte(s)\n",  sizeof(d));
+	printf("Size of a long long int: %ln byte(s)\n",  sizeof(g));
 	return (0);
 }
